minor_changes

+++ srcs/rays/rays_setup.c
@@ -6,17 +6,17 @@
 /*   By: mbos <mbos@student.le-101.fr>              +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2019/12/26 14:24:41 by mbos         #+#   ##    ##    #+#       */
-/*   Updated: 2020/01/04 14:32:32 by mbos        ###    #+. /#+    ###.fr     */
+/*   Updated: 2020/01/06 16:44:18 by mbos        ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
 
 #include "../../includes/cube3d.h"
 
-t_rays	ray_setup(double angle)
+t_rays	ray_setup(float angle)
 {
 	t_rays	ray;
-	double	angle_radian;
+	float	angle_radian;
 
 	angle_radian = degrees_to_radian(angle);
 	ray = rays_init();
