+++ srcs/utils/angles.c
@@ -1,19 +1,19 @@
 /* ************************************************************************** */
 /*                                                          LE - /            */
 /*                                                              /             */
-/*   maths.c                                          .::    .:/ .      .::   */
+/*   angles.c                                         .::    .:/ .      .::   */
 /*                                                 +:+:+   +:    +:  +:+:+    */
 /*   By: mbos <mbos@student.le-101.fr>              +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2019/12/26 14:25:03 by mbos         #+#   ##    ##    #+#       */
-/*   Updated: 2019/12/26 17:21:48 by mbos        ###    #+. /#+    ###.fr     */
+/*   Updated: 2020/01/06 16:45:01 by mbos        ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
 
 #include "../../includes/cube3d.h"
 
-double     degrees_to_radian(double angle_in_degrees)
+float     degrees_to_radian(float angle_in_degrees)
 {
     return (angle_in_degrees * (M_PI / 180));
 }
